backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-socketio==5.3.3
python-vlc==3.0.18120
mutagen==1.46.0
sqlalchemy==2.0.7
pytest==7.3.1
requests==2.28.2
marshmallow==3.19.0
python-engineio==4.4.1
python-socketio==5.8.0
eventlet==0.33.3
gevent==22.10.2
